requirements.txt
pip==24.0
pandas==1.3.4
sklearn==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
streamlit-option-menu==0.3.2
